code reformat (new)

+++ modApiServer/test/org/aion/api/server/ApiAionTest.java
@@ -23,7 +23,21 @@
 
 package org.aion.api.server;
 
+import static org.junit.Assert.assertArrayEquals;
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNotEquals;
+import static org.junit.Assert.assertNotNull;
+import static org.junit.Assert.assertNull;
+import static org.junit.Assert.assertTrue;
+
 import io.undertow.util.FileUtils;
+import java.io.File;
+import java.io.IOException;
+import java.math.BigInteger;
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.List;
+import java.util.Map;
 import org.aion.api.server.rpc.ApiWeb3Aion;
 import org.aion.api.server.rpc.RpcError;
 import org.aion.api.server.rpc.RpcMsg;
@@ -54,16 +68,6 @@ import org.junit.Before;
 import org.junit.Ignore;
 import org.junit.Test;
 
-import java.io.File;
-import java.io.IOException;
-import java.math.BigInteger;
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
-import java.util.Map;
-
-import static org.junit.Assert.*;
-
 public class ApiAionTest {
 
     private class ApiAionImpl extends ApiAion {
@@ -90,7 +94,6 @@ public class ApiAionTest {
         private boolean allFlagsSet() {
             return (onBlockFlag && pendingRcvdFlag && pendingUpdateFlag);
         }
-
 
         private ApiAionImpl(AionImpl impl) {
             super(impl);
@@ -124,11 +127,10 @@ public class ApiAionTest {
 
             ees.add(evBlock);
 
-            //provokes exception in EpApi.run()
+            // provokes exception in EpApi.run()
             ees.add(new EventBlock(EventBlock.CALLBACK.ONBLOCK0));
             ees.add(new EventDummy());
         }
-
     }
 
     private static final String KEYSTORE_PATH;
@@ -163,23 +165,19 @@ public class ApiAionTest {
         // get a list of all the files in keystore directory
         File folder = new File(KEYSTORE_PATH);
 
-        if (folder == null)
-            return;
+        if (folder == null) return;
 
         File[] AllFilesInDirectory = folder.listFiles();
 
         // check for invalid or wrong path - should not happen
-        if (AllFilesInDirectory == null)
-            return;
+        if (AllFilesInDirectory == null) return;
 
         for (File file : AllFilesInDirectory) {
-            if (file.lastModified() >= testStartTime)
-                file.delete();
+            if (file.lastModified() >= testStartTime) file.delete();
         }
         folder = new File(DATABASE_PATH);
 
-        if (folder == null)
-            return;
+        if (folder == null) return;
 
         try {
             FileUtils.deleteRecursive(folder.toPath());
@@ -236,13 +234,16 @@ public class ApiAionTest {
         assertTrue(rslt.getKey().isEqual(blk));
 
         // check because blk might be the genesis block
-        assertEquals(rslt.getValue(),
-                ((AionBlockStore) impl.getBlockchain().getBlockStore()).getTotalDifficultyForHash(blk.getHash()));
+        assertEquals(
+                rslt.getValue(),
+                ((AionBlockStore) impl.getBlockchain().getBlockStore())
+                        .getTotalDifficultyForHash(blk.getHash()));
 
         // retrieving genesis block's difficulty
-        assertEquals(api.getBlockWithTotalDifficulty(0).getValue(), CfgAion.inst().getGenesis().getDifficultyBI());
+        assertEquals(
+                api.getBlockWithTotalDifficulty(0).getValue(),
+                CfgAion.inst().getGenesis().getDifficultyBI());
     }
-
 
     @Test
     public void testGetSync() {
@@ -250,30 +251,37 @@ public class ApiAionTest {
         assertNotNull(sync);
         assertEquals(sync.done, impl.isSyncComplete());
         if (impl.getInitialStartingBlockNumber().isPresent())
-            assertEquals((long) impl.getInitialStartingBlockNumber().get(), sync.chainStartingBlkNumber);
-        else
-            assertEquals(0L, sync.chainStartingBlkNumber);
+            assertEquals(
+                    (long) impl.getInitialStartingBlockNumber().get(), sync.chainStartingBlkNumber);
+        else assertEquals(0L, sync.chainStartingBlkNumber);
         if (impl.getNetworkBestBlockNumber().isPresent())
             assertEquals((long) impl.getNetworkBestBlockNumber().get(), sync.networkBestBlkNumber);
-        else
-            assertEquals(0L, sync.networkBestBlkNumber);
+        else assertEquals(0L, sync.networkBestBlkNumber);
         if (impl.getLocalBestBlockNumber().isPresent())
             assertEquals((long) impl.getLocalBestBlockNumber().get(), sync.chainBestBlkNumber);
-        else
-            assertEquals(0L, sync.chainBestBlkNumber);
+        else assertEquals(0L, sync.chainBestBlkNumber);
     }
 
-    @Test @Ignore
+    @Test
+    @Ignore
     public void testGetTransactions() {
         AionBlock parentBlk = impl.getBlockchain().getBestBlock();
         byte[] msg = "test message".getBytes();
-        AionTransaction tx = new AionTransaction(repo.getNonce(Address.ZERO_ADDRESS()).toByteArray(),
-                Address.ZERO_ADDRESS(), Address.ZERO_ADDRESS(), BigInteger.ONE.toByteArray(),
-                msg, 100000, 100000);
+        AionTransaction tx =
+                new AionTransaction(
+                        repo.getNonce(Address.ZERO_ADDRESS()).toByteArray(),
+                        Address.ZERO_ADDRESS(),
+                        Address.ZERO_ADDRESS(),
+                        BigInteger.ONE.toByteArray(),
+                        msg,
+                        100000,
+                        100000);
         tx.sign(new ECKeyEd25519());
 
-        AionBlock blk = impl.getAionHub().getBlockchain().createNewBlock(parentBlk,
-                Collections.singletonList(tx), false);
+        AionBlock blk =
+                impl.getAionHub()
+                        .getBlockchain()
+                        .createNewBlock(parentBlk, Collections.singletonList(tx), false);
 
         assertNotNull(blk);
         assertNotEquals(blk.getTransactionsList().size(), 0);
@@ -288,12 +296,13 @@ public class ApiAionTest {
 
         blk = api.getBlockByHash(blk.getHash());
 
-        assertEquals(1, api.getTransactionCount(
-                blk.getTransactionsList().get(0).getFrom(), blk.getNumber()));
+        assertEquals(
+                1,
+                api.getTransactionCount(
+                        blk.getTransactionsList().get(0).getFrom(), blk.getNumber()));
         assertEquals(0, api.getTransactionCount(Address.EMPTY_ADDRESS(), blk.getNumber()));
 
         assertEquals(tx, api.getTransactionByHash(tx.getHash()));
-
     }
 
     @Test
@@ -303,14 +312,26 @@ public class ApiAionTest {
         Address addr = new Address(Keystore.create("testPwd"));
         AccountManager.inst().unlockAccount(addr, "testPwd", 50000);
 
-        AionTransaction tx = new AionTransaction(repo.getNonce(Address.ZERO_ADDRESS()).toByteArray(),
-                addr, Address.ZERO_ADDRESS(), BigInteger.ONE.toByteArray(),
-                msg, 100000, 100000);
+        AionTransaction tx =
+                new AionTransaction(
+                        repo.getNonce(Address.ZERO_ADDRESS()).toByteArray(),
+                        addr,
+                        Address.ZERO_ADDRESS(),
+                        BigInteger.ONE.toByteArray(),
+                        msg,
+                        100000,
+                        100000);
         tx.sign(new ECKeyEd25519());
 
-
-        ArgTxCall txcall = new ArgTxCall(addr, Address.ZERO_ADDRESS(),
-                msg, repo.getNonce(addr), BigInteger.ONE, 100000, 100000);
+        ArgTxCall txcall =
+                new ArgTxCall(
+                        addr,
+                        Address.ZERO_ADDRESS(),
+                        msg,
+                        repo.getNonce(addr),
+                        BigInteger.ONE,
+                        100000,
+                        100000);
 
         assertNotNull(api.doCall(txcall));
         tearDown();
@@ -324,14 +345,26 @@ public class ApiAionTest {
 
         AccountManager.inst().unlockAccount(addr, "testPwd", 50000);
 
-        AionTransaction tx = new AionTransaction(repo.getNonce(Address.ZERO_ADDRESS()).toByteArray(),
-                addr, Address.ZERO_ADDRESS(), BigInteger.ONE.toByteArray(),
-                msg, 100000, 100000);
+        AionTransaction tx =
+                new AionTransaction(
+                        repo.getNonce(Address.ZERO_ADDRESS()).toByteArray(),
+                        addr,
+                        Address.ZERO_ADDRESS(),
+                        BigInteger.ONE.toByteArray(),
+                        msg,
+                        100000,
+                        100000);
         tx.sign(new ECKeyEd25519());
 
-
-        ArgTxCall txcall = new ArgTxCall(addr, Address.ZERO_ADDRESS(),
-                msg, repo.getNonce(addr), BigInteger.ONE, 100000, 100000);
+        ArgTxCall txcall =
+                new ArgTxCall(
+                        addr,
+                        Address.ZERO_ADDRESS(),
+                        msg,
+                        repo.getNonce(addr),
+                        BigInteger.ONE,
+                        100000,
+                        100000);
 
         assertEquals(impl.estimateTxNrg(tx, api.getBestBlock()), api.estimateNrg(txcall));
         tearDown();
@@ -345,19 +378,40 @@ public class ApiAionTest {
 
         AccountManager.inst().unlockAccount(addr, "testPwd", 50000);
 
-        ArgTxCall txcall = new ArgTxCall(addr, Address.ZERO_ADDRESS(),
-                msg, repo.getNonce(addr), BigInteger.ONE, 100000, 100000);
+        ArgTxCall txcall =
+                new ArgTxCall(
+                        addr,
+                        Address.ZERO_ADDRESS(),
+                        msg,
+                        repo.getNonce(addr),
+                        BigInteger.ONE,
+                        100000,
+                        100000);
 
         assertNotNull(api.createContract(txcall).transId);
         assertNotNull(api.createContract(txcall).address);
 
-        txcall = new ArgTxCall(null, Address.ZERO_ADDRESS(),
-                msg, repo.getNonce(addr), BigInteger.ONE, 100000, 100000);
+        txcall =
+                new ArgTxCall(
+                        null,
+                        Address.ZERO_ADDRESS(),
+                        msg,
+                        repo.getNonce(addr),
+                        BigInteger.ONE,
+                        100000,
+                        100000);
 
         assertNull(api.createContract(txcall));
 
-        txcall = new ArgTxCall(Address.ZERO_ADDRESS(), Address.ZERO_ADDRESS(),
-                msg, repo.getNonce(addr), BigInteger.ONE, 100000, 100000);
+        txcall =
+                new ArgTxCall(
+                        Address.ZERO_ADDRESS(),
+                        Address.ZERO_ADDRESS(),
+                        msg,
+                        repo.getNonce(addr),
+                        BigInteger.ONE,
+                        100000,
+                        100000);
 
         assertNull(api.createContract(txcall));
         tearDown();
@@ -365,10 +419,15 @@ public class ApiAionTest {
 
     @Test
     public void testAccountGetters() {
-        assertEquals(repo.getBalance(Address.ZERO_ADDRESS()), api.getBalance(Address.ZERO_ADDRESS()));
+        assertEquals(
+                repo.getBalance(Address.ZERO_ADDRESS()), api.getBalance(Address.ZERO_ADDRESS()));
         assertEquals(repo.getNonce(Address.ZERO_ADDRESS()), api.getNonce(Address.ZERO_ADDRESS()));
-        assertEquals(repo.getBalance(Address.ZERO_ADDRESS()), api.getBalance(Address.ZERO_ADDRESS().toString()));
-        assertEquals(repo.getNonce(Address.ZERO_ADDRESS()), api.getNonce(Address.ZERO_ADDRESS().toString()));
+        assertEquals(
+                repo.getBalance(Address.ZERO_ADDRESS()),
+                api.getBalance(Address.ZERO_ADDRESS().toString()));
+        assertEquals(
+                repo.getNonce(Address.ZERO_ADDRESS()),
+                api.getNonce(Address.ZERO_ADDRESS().toString()));
     }
 
     @Test
@@ -379,25 +438,52 @@ public class ApiAionTest {
 
         AccountManager.inst().unlockAccount(addr, "testPwd", 50000);
 
-        ArgTxCall txcall = new ArgTxCall(addr, Address.ZERO_ADDRESS(),
-                msg, repo.getNonce(addr), BigInteger.ONE, 100000, 100000);
+        ArgTxCall txcall =
+                new ArgTxCall(
+                        addr,
+                        Address.ZERO_ADDRESS(),
+                        msg,
+                        repo.getNonce(addr),
+                        BigInteger.ONE,
+                        100000,
+                        100000);
 
         assertNotNull(api.sendTransaction(txcall));
 
-        AionTransaction tx = new AionTransaction(repo.getNonce(Address.ZERO_ADDRESS()).toByteArray(),
-                addr, Address.ZERO_ADDRESS(), BigInteger.ONE.toByteArray(),
-                msg, 100000, 100000);
+        AionTransaction tx =
+                new AionTransaction(
+                        repo.getNonce(Address.ZERO_ADDRESS()).toByteArray(),
+                        addr,
+                        Address.ZERO_ADDRESS(),
+                        BigInteger.ONE.toByteArray(),
+                        msg,
+                        100000,
+                        100000);
         tx.sign(new ECKeyEd25519());
 
         assertNotNull(api.sendTransaction(tx.getEncoded()));
 
-        txcall = new ArgTxCall(null, Address.ZERO_ADDRESS(),
-                msg, repo.getNonce(addr), BigInteger.ONE, 100000, 100000);
+        txcall =
+                new ArgTxCall(
+                        null,
+                        Address.ZERO_ADDRESS(),
+                        msg,
+                        repo.getNonce(addr),
+                        BigInteger.ONE,
+                        100000,
+                        100000);
 
         assertNull(api.sendTransaction(txcall));
 
-        txcall = new ArgTxCall(Address.EMPTY_ADDRESS(), Address.ZERO_ADDRESS(),
-                msg, repo.getNonce(addr), BigInteger.ONE, 100000, 100000);
+        txcall =
+                new ArgTxCall(
+                        Address.EMPTY_ADDRESS(),
+                        Address.ZERO_ADDRESS(),
+                        msg,
+                        repo.getNonce(addr),
+                        BigInteger.ONE,
+                        100000,
+                        100000);
 
         assertNull(api.sendTransaction(txcall));
         tearDown();
@@ -405,15 +491,14 @@ public class ApiAionTest {
 
     @Test
     public void testSimpleGetters() {
-        assertEquals(CfgAion.inst().getApi().getNrg().getNrgPriceDefault(),
+        assertEquals(
+                CfgAion.inst().getApi().getNrg().getNrgPriceDefault(),
                 api.getRecommendedNrgPrice());
         api.initNrgOracle(impl);
-        assertEquals(api.getNrgOracle().getNrgPrice(),
-                api.getRecommendedNrgPrice());
+        assertEquals(api.getNrgOracle().getNrgPrice(), api.getRecommendedNrgPrice());
 
         assertNotNull(api.getCoinbase());
-        assertEquals(repo.getCode(Address.ZERO_ADDRESS()),
-                api.getCode(Address.ZERO_ADDRESS()));
+        assertEquals(repo.getCode(Address.ZERO_ADDRESS()), api.getCode(Address.ZERO_ADDRESS()));
         assertEquals(impl.getBlockMiner().isMining(), api.isMining());
         assertArrayEquals(CfgAion.inst().getNodes(), api.getBootNodes());
         assertEquals(impl.getAionHub().getP2pMgr().getActiveNodes().size(), api.peerCount());
@@ -427,7 +512,10 @@ public class ApiAionTest {
         double hashRate = 1000;
 
         assertTrue(api.setReportedHashrate(Double.toString(hashRate), ""));
-        assertEquals(impl.getBlockMiner().getHashrate() + hashRate, Double.parseDouble(api.getHashrate()), 0.001);
+        assertEquals(
+                impl.getBlockMiner().getHashrate() + hashRate,
+                Double.parseDouble(api.getHashrate()),
+                0.001);
     }
 
     @Test
@@ -462,16 +550,21 @@ public class ApiAionTest {
         assertNotNull(tx);
 
         assertEquals(tx.get("to"), outTx.get("to"));
-        assertEquals(tx.get("value").toString(),
-            TypeConverter.StringHexToBigInteger(outTx.get("value").toString()).toString());
-        assertEquals(tx.get("gasPrice").toString(),
-            TypeConverter.StringHexToBigInteger(outTx.get("gasPrice").toString()).toString());
-        assertEquals(tx.get("gasPrice").toString(),
-            TypeConverter.StringHexToBigInteger(outTx.get("nrgPrice").toString()).toString());
-        assertEquals(tx.get("gas").toString(),
-            TypeConverter.StringHexToBigInteger(outTx.get("gas").toString()).toString());
-        assertEquals(tx.get("gas").toString(),
-            TypeConverter.StringHexToBigInteger(outTx.get("nrg").toString()).toString());
+        assertEquals(
+                tx.get("value").toString(),
+                TypeConverter.StringHexToBigInteger(outTx.get("value").toString()).toString());
+        assertEquals(
+                tx.get("gasPrice").toString(),
+                TypeConverter.StringHexToBigInteger(outTx.get("gasPrice").toString()).toString());
+        assertEquals(
+                tx.get("gasPrice").toString(),
+                TypeConverter.StringHexToBigInteger(outTx.get("nrgPrice").toString()).toString());
+        assertEquals(
+                tx.get("gas").toString(),
+                TypeConverter.StringHexToBigInteger(outTx.get("gas").toString()).toString());
+        assertEquals(
+                tx.get("gas").toString(),
+                TypeConverter.StringHexToBigInteger(outTx.get("nrg").toString()).toString());
         assertEquals("0x", outTx.get("input").toString());
 
         JSONArray rawTxArray = new JSONArray();
@@ -497,7 +590,7 @@ public class ApiAionTest {
 
         JSONArray jsonArray = new JSONArray();
         jsonArray.put(tx);
-        //don't pass address
+        // don't pass address
 
         RpcMsg rpcMsg = web3Api.eth_signTransaction(jsonArray);
         assertNotNull(rpcMsg);
